+++ miv/signal/filter/butter_bandpass_filter.py
@@ -13,7 +13,7 @@ import numpy.typing as npt
 import scipy.signal as sps
 
 from miv.core.datatype import Signal
-from miv.core.operator import GeneratorOperatorMixin
+from miv.core.operator_generator import GeneratorOperatorMixin
 from miv.core.wrapper import cache_generator_call
 from miv.typing import SignalType
